Add handling of network state and a banner to notify the user about it

+++ app/src/main/java/com/app/realestatedemoapp/data/repository/PropertyRepositoryImpl.kt
@@ -9,15 +9,20 @@ import com.app.realestatedemoapp.data.local.dao.PropertyDao
 import com.app.realestatedemoapp.data.mapper.toDomain
 import com.app.realestatedemoapp.data.mapper.toEntity
 import com.app.realestatedemoapp.data.remote.ApiService
+import com.app.realestatedemoapp.domain.NetworkConnectivityObserver
 import com.app.realestatedemoapp.domain.PropertyRepository
+import com.app.realestatedemoapp.domain.model.NetworkStatus
 import com.app.realestatedemoapp.domain.model.PropertyModel
 import com.app.realestatedemoapp.util.Constants
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
 
 class PropertyRepositoryImpl @Inject constructor(
-    private val apiService: ApiService, private val propertyDao: PropertyDao
+    private val apiService: ApiService,
+    private val propertyDao: PropertyDao,
+    val connectivityObserver: NetworkConnectivityObserver
 ) : PropertyRepository {
 
     @OptIn(ExperimentalPagingApi::class)
@@ -34,6 +39,9 @@ class PropertyRepositoryImpl @Inject constructor(
     }
 
     override suspend fun refreshProperties() {
+        if (connectivityObserver.observe().first() == NetworkStatus.Disconnected) {
+            return
+        }
         val properties = apiService.getProperties()
         val bookmarkedIds = propertyDao.getBookmarkedProperties().map { it.id }.toSet()
         propertyDao.deleteAllProperties()
